# Improve perf of base64 encoding check on a consumer hot path